The system has been changed so that individual statistics pages can be served according to the user's login information. Instead, we had to stop using getStaticProps and revert to getServerSideProps, which again caused performance problems.
