+++ components/ProfileList.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { doc, getDoc } from "firebase/firestore";
-import { db } from "../config/firebase";
+import { db } from "../config/firebaseClient";
 
 const ProfileList = () => {
   const [profile, setProfile] = useState({
