+++ package.json
@@ -21,6 +21,7 @@
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^27.3.1",
     "chart.js": "^3.6.0",
+    "dotenv": "^16.0.0",
     "firebase": "^9.6.6",
     "firebase-admin": "^10.0.2",
     "flowbite": "^1.3.3",
@@ -31,6 +32,7 @@
     "react-dom": "17.0.2",
     "react-firebase-hooks": "^3.0.4",
     "react-loading": "^2.0.3",
+    "sharp": "^0.30.3",
     "styled-components": "^5.3.3",
     "swr": "^1.2.0"
   },
