+++ pages/_app.tsx
@@ -1,10 +1,10 @@
+import type { AppProps } from "next/app";
 import { AuthProvider } from "../auth";
 import Layout from "../components/Layout";
 import "../styles/globals.css";
 // import Script from "next/script";
 
-// @ts-ignore
-function MyApp({ Component, pageProps }) {
+function MyApp({ Component, pageProps }: AppProps) {
   return (
     <Layout>
       <AuthProvider>
