+++ pages/index.tsx
@@ -1,12 +1,14 @@
 // Next.js and React related
 import Head from 'next/head';
 import { useEffect } from 'react';
-import { GetStaticProps } from 'next';
+import { GetServerSideProps, GetServerSidePropsContext } from 'next';
+import nookies from 'nookies';
 // import { useState } from "react";
 
 // Config
-// import { auth } from "../config/firebase";
+import { auth } from '../config/firebaseClient';
 // import { useAuth } from "../auth";
+import { verifyIdToken } from '../firebaseAdmin';
 
 // Components
 import ProfileList from '../components/ProfileList';
@@ -63,14 +65,6 @@ export default function Home({
   //   setOpen(false);
   // };
 
-  // useEffect(() => {
-  //   import("flowbite");
-  //   import('@themesberg/flowbite');
-  // }, []);
-
-  // console.log(`Propped data is: ${JSON.stringify(data)}`);
-  // console.log(`Propped data is: ${data.messages.total}`);
-
   return (
     <div>
       <Head>
@@ -114,80 +108,98 @@ export default function Home({
 
 // This gets called on every request.
 // The official document is here: https://nextjs.org/docs/basic-features/data-fetching/get-server-side-props
-export const getStaticProps: GetStaticProps = async () => {
-  // The docID should be changed to get it when clicking on the list of transition sources, or if not, get it from the firebase login user.
-  const docID = 'REArvdg1hv5I6pkJ40nC';
-  const userDoc = await getAUserDoc(docID);
+export const getServerSideProps: GetServerSideProps = async (ctx: GetServerSidePropsContext) => {
+  // export const getStaticProps: GetStaticProps = async () => {
+  try {
+    // The docID should be changed to get it when clicking on the list of transition sources, or if not, get it from the firebase login user.
+    const cookies = nookies.get(ctx);
+    const token = await verifyIdToken(cookies.token);
 
-  // Parameters for asana
-  const asanaPersonalAccessToken =
-    userDoc?.asana?.workspace[0].personalAccessToken;
-  const asanaUserId: string | undefined = userDoc?.asana?.userId;
-  const asanaWorkspaceId: string | undefined =
-    userDoc?.asana?.workspace[0].workspaceId;
+    // the user is authenticated!
+    const { uid, email } = token;
+    // const currentUser = auth.currentUser;
+    // const docId = currentUser?.uid ? currentUser.uid : '';
+    const userDoc = await getAUserDoc(uid);
 
-  // Parameters for github
-  const githubRepoName: string | undefined =
-    userDoc?.github?.repositories[0].repo;
-  const githubOwnerName: string | undefined =
-    userDoc?.github?.repositories[0].owner;
-  const githubUserId: number | undefined = userDoc?.github?.userId;
-  const githubUserName: string | undefined = userDoc?.github?.userName;
+    // Parameters for asana
+    const asanaPersonalAccessToken =
+      userDoc?.asana?.workspace[0].personalAccessToken;
+    const asanaUserId: string | undefined = userDoc?.asana?.userId;
+    const asanaWorkspaceId: string | undefined =
+      userDoc?.asana?.workspace[0].workspaceId;
 
-  // Parameters for slack
-  const searchQuery: string | undefined = userDoc?.slack?.workspace[0].memberId;
-  const slackUserToken: string = `Bearer ${userDoc?.slack?.workspace[0].userToken}`;
+    // Parameters for github
+    const githubRepoName: string | undefined =
+      userDoc?.github?.repositories[0].repo;
+    const githubOwnerName: string | undefined =
+      userDoc?.github?.repositories[0].owner;
+    const githubUserId: number | undefined = userDoc?.github?.userId;
+    const githubUserName: string | undefined = userDoc?.github?.userName;
 
-  // Tabulate number of times a user has been mentioned in all slack public channels
-  const numberOfMentioned = await slackSearchFromServer(
-    // @ts-ignore
-    searchQuery,
-    slackUserToken,
-  );
+    // Parameters for slack
+    const searchQuery: string | undefined =
+      userDoc?.slack?.workspace[0].memberId;
+    const slackUserToken: string = `Bearer ${userDoc?.slack?.workspace[0].userToken}`;
 
-  // Tabulate number of times a user has newly sent messages in all slack public channels
-  const channelList = await slackConversationList(slackUserToken);
-  let numberOfNewSent: number = 0;
-  for (let x in channelList) {
-    let channel = channelList[x];
-    numberOfNewSent += await slackConversationHistory(
-      channel,
+    // Tabulate number of times a user has been mentioned in all slack public channels
+    const numberOfMentioned = await slackSearchFromServer(
       // @ts-ignore
       searchQuery,
       slackUserToken,
     );
-  }
 
-  // Tabulate number of times a user has replied in all slack public channels
-  let numberOfReplies: number = 0;
-  for (let x in channelList) {
-    // console.log(`x is: ${x} times`);
-    let channel = channelList[x];
-    // console.log(`Current channel is: ${channelList[x]}`);
-    let timestampList = await listTimestampInSlack(channel, slackUserToken);
-    // console.log(`Timestamp list length is: ${timestampList.length}`);
-    // console.log(`Timestamp list is: ${timestampList}`);
-    for (let y in timestampList) {
-      // console.log(`y is: ${y} times`);
-      // @ts-ignore
-      numberOfReplies += await countRepliesInSlack(channel, timestampList[y], searchQuery,slackUserToken);
+    // Tabulate number of times a user has newly sent messages in all slack public channels
+    const channelList = await slackConversationList(slackUserToken);
+    let numberOfNewSent: number = 0;
+    for (let x in channelList) {
+      let channel = channelList[x];
+      numberOfNewSent += await slackConversationHistory(
+        channel,
+        // @ts-ignore
+        searchQuery,
+        slackUserToken,
+      );
     }
-  }
-  // console.log(`numberOfReplies is: ${numberOfReplies}`);
 
-  // Pass data to the page via props
-  return {
-    props: {
-      numberOfMentioned,
-      numberOfNewSent,
-      numberOfReplies,
-      asanaUserId,
-      asanaWorkspaceId,
-      asanaPersonalAccessToken,
-      githubRepoName,
-      githubOwnerName,
-      githubUserId,
-      githubUserName,
-    },
-  };
+    // Tabulate number of times a user has replied in all slack public channels
+    let numberOfReplies: number = 0;
+    for (let x in channelList) {
+      let channel = channelList[x];
+      let timestampList = await listTimestampInSlack(channel, slackUserToken);
+      for (let y in timestampList) {
+        numberOfReplies += await countRepliesInSlack(
+          channel,
+          timestampList[y],
+          // @ts-ignore
+          searchQuery,
+          slackUserToken,
+        );
+      }
+    }
+
+    // Pass data to the page via props
+    return {
+      props: {
+        numberOfMentioned,
+        numberOfNewSent,
+        numberOfReplies,
+        asanaUserId,
+        asanaWorkspaceId,
+        asanaPersonalAccessToken,
+        githubRepoName,
+        githubOwnerName,
+        githubUserId,
+        githubUserName,
+      },
+    };
+  } catch (error) {
+    // either the `token` cookie didn't exist or token verification failed.
+    // either way: redirect to the login page
+    ctx.res.writeHead(302, { Location: '/login' });
+    ctx.res.end();
+
+    // `as never` prevents inference issues with InferGetServerSidePropsType.
+    // The props returned here don't matter because we've already redirected the user.
+    return { props: {} as never };
+  }
 };
