+++ pages/login.tsx
@@ -1,0 +1,4 @@
+const login = () => {
+  return <div>login page here</div>;
+};
+export default login;
