+++ tsconfig.json
@@ -26,7 +26,8 @@
   "include": [
     "next-env.d.ts",
     "**/*.ts",
-    "**/*.tsx"
+    "**/*.tsx", 
+    "config/firebaseClient.js"
   ],
   "exclude": [
     "node_modules"
